+++ experimental/packages/opentelemetry-sdk-metrics-base/src/MeterProvider.ts
@@ -19,15 +19,9 @@ import * as metrics from '@opentelemetry/api-metrics';
 import { Resource } from '@opentelemetry/resources';
 import { MetricReader } from './export/MetricReader';
 import { MeterProviderSharedState } from './state/MeterProviderSharedState';
-import { InstrumentSelector } from './view/InstrumentSelector';
-import { MeterSelector } from './view/MeterSelector';
-import { View } from './view/View';
 import { MetricCollector } from './state/MetricCollector';
-import { Aggregation } from './view/Aggregation';
-import { FilteringAttributesProcessor } from './view/AttributesProcessor';
-import { InstrumentType } from './InstrumentDescriptor';
-import { PatternPredicate } from './view/Predicate';
 import { ForceFlushOptions, ShutdownOptions } from './types';
+import { View } from './view/View';
 
 /**
  * MeterProviderOptions provides an interface for configuring a MeterProvider.
@@ -35,68 +29,7 @@ import { ForceFlushOptions, ShutdownOptions } from './types';
 export interface MeterProviderOptions {
   /** Resource associated with metric telemetry  */
   resource?: Resource;
-}
-
-export type ViewOptions = {
-  /**
-   *  If not provided, the Instrument name will be used by default. This will be used as the name of the metrics stream.
-   */
-  name?: string,
-  /**
-   * If not provided, the Instrument description will be used by default.
-   */
-  description?: string,
-  /**
-   * If provided, the attributes that are not in the list will be ignored.
-   * If not provided, all the attribute keys will be used by default.
-   */
-  attributeKeys?: string[],
-  /**
-   * The {@link Aggregation} aggregation to be used.
-   */
-  aggregation?: Aggregation,
-
-  // TODO: Add ExemplarReservoir
-};
-
-export type SelectorOptions = {
-  /**
-   * Selects instrument by their fields.
-   */
-  instrument?: {
-    /**
-     * The type of the Instrument(s).
-     */
-    type?: InstrumentType,
-    /**
-     * Name of the Instrument(s) with wildcard support.
-     */
-    name?: string,
-  }
-  /**
-   * Selects instrument by fields of their meter.
-   */
-  meter?: {
-    /**
-     * The name of the Meter.
-     */
-    name?: string;
-    /**
-     * The version of the Meter.
-     */
-    version?: string;
-    /**
-     * The schema URL of the Meter.
-     */
-    schemaUrl?: string;
-  }
-};
-
-function isViewOptionsEmpty(options: ViewOptions): boolean {
-  return (options.name == null &&
-    options.aggregation == null &&
-    options.attributeKeys == null &&
-    options.description == null);
+  views?: View[];
 }
 
 /**
@@ -108,6 +41,11 @@ export class MeterProvider implements metrics.MeterProvider {
 
   constructor(options?: MeterProviderOptions) {
     this._sharedState = new MeterProviderSharedState(options?.resource ?? Resource.empty());
+    if(options?.views != null && options.views.length > 0){
+      for(const view of options.views){
+        this._sharedState.viewRegistry.addView(view);
+      }
+    }
   }
 
   /**
@@ -135,37 +73,6 @@ export class MeterProvider implements metrics.MeterProvider {
     const collector = new MetricCollector(this._sharedState, metricReader);
     metricReader.setMetricProducer(collector);
     this._sharedState.metricCollectors.push(collector);
-  }
-
-  addView(options: ViewOptions, selectorOptions?: SelectorOptions) {
-    if (isViewOptionsEmpty(options)) {
-      throw new Error('Cannot create view with no view arguments supplied');
-    }
-
-    // the SDK SHOULD NOT allow Views with a specified name to be declared with instrument selectors that
-    // may select more than one instrument (e.g. wild card instrument name) in the same Meter.
-    if (options.name != null &&
-      (selectorOptions?.instrument?.name == null ||
-        PatternPredicate.hasWildcard(selectorOptions.instrument.name))) {
-      throw new Error('Views with a specified name must be declared with an instrument selector that selects at most one instrument per meter.');
-    }
-
-    // Create AttributesProcessor if attributeKeys are defined set.
-    let attributesProcessor = undefined;
-    if (options.attributeKeys != null) {
-      attributesProcessor = new FilteringAttributesProcessor(options.attributeKeys);
-    }
-
-    const view = new View({
-      name: options.name,
-      description: options.description,
-      aggregation: options.aggregation,
-      attributesProcessor: attributesProcessor
-    });
-    const instrument = new InstrumentSelector(selectorOptions?.instrument);
-    const meter = new MeterSelector(selectorOptions?.meter);
-
-    this._sharedState.viewRegistry.addView(view, instrument, meter);
   }
 
   /**